feat(0): [wcf-bmc]-修改文件下载接口为文件保存，实际是将文件保存至本服务指定目录

+++ clients/java/wcf-bmc/src/main/java/com/wechat/ferry/handle/WeChatSocketClient.java
@@ -639,7 +639,7 @@ public class WeChatSocketClient {
      * @param id 待撤回消息的 id
      * @return 结果状态码 0 为成功，其他失败
      */
-    public int revokeMsg(Integer id) {
+    public int revokeMsg(Long id) {
         int ret = -1;
         Request req = Request.newBuilder().setFuncValue(Functions.FUNC_REVOKE_MSG_VALUE).setUi64(id).build();
         Response rsp = sendCmd(req);
